% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_preprocess_plate_data_new.R
\name{convert_timestamp}
\alias{convert_timestamp}
\title{Convert the the time column in the WAVE input dataframe to a time scale in seconds}
\usage{
convert_timestamp(plate_df)
}
\arguments{
\item{plate_df}{A dataframe generated in adjust_columns}
}
\value{
A new dataframe with new columns added  to \code{plate_df}. New columns 
 are: "totalMinutes", "minutes", "timescale"
}
\description{
Convert the the time column in the WAVE input dataframe to a time scale in seconds
}
