% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{get_injection_info_H}
\alias{get_injection_info_H}
\title{Read the injection information from the "Operation log" sheet H-version}
\usage{
get_injection_info_H(filePathSeahorse)
}
\arguments{
\item{filePathSeahorse}{Absolute path to the Seahorse excel data derived from the Agilent Seahorse XF Wave software. assay result file (.asyr) files must converted into .excel files.}
}
\value{
A new dataframe called measurement_info The df has three columns: 
  $measurement, $interval, $injection
}
\description{
IMPORTANT: the output of the injection information is different on different XFe96 instruments. Two version of
  this function are available. The get_injection_info_H assumes the names of the injections are listed in the
  "operation log file". The other (get_injection_info_M) function uses a manual assignment of the injection names.
  In both cases, the number of measurements per injection ARE read by the function
}
\keyword{internal}
