% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{get_injection_info_M}
\alias{get_injection_info_M}
\title{Read the injection information from the "Operation log" sheet M-version}
\usage{
get_injection_info_M(filePathSeahorse)
}
\arguments{
\item{filePathSeahorse}{The name of the file that is being processed}
}
\value{
A new dataframe called measurement_info The df has three columns: 
  $measurement, $interval, $injection
}
\description{
IMPORTANT: the output of the injection information is different on different XFe96 instruments. Two version of
  this function are available. The get_injection_info_H assumes the names of the injections are listed in the
  "operation log file". The other (get_injection_info_M) function uses a manual assignment of the injection names.
  In both cases, the number of measurements per injection ARE read by the function.
  The default names for the get_injection_info_M are c("basal", "OM", "FCCP", "AM/rot").
}
\examples{
get_injection_info_M(filePathSeahorse)
}
