% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{get_originalRateTable}
\alias{get_originalRateTable}
\title{Title:}
\usage{
get_originalRateTable(filePathSeahorse)
}
\arguments{
\item{filePathSeahorse}{Absolute path to the Seahorse excel data derived from the Agilent Seahorse XF Wave software. assay result file (.asyr) files must converted into .excel files.}
}
\value{
original_rate_df_list
}
\description{
Title:
}
\keyword{internal}
