% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_QC_module_forShiny.R
\name{get_plate_scores}
\alias{get_plate_scores}
\title{Getting plate scores}
\usage{
get_plate_scores(df, qc_plate)
}
\arguments{
\item{df}{Stat summaries for NEW PLATES - well (note : this is a tibble in list)

A tibble: 1 × 8
plate_id       well Maximum Minimum First  Last Range(M-M) Range(F-L)
<chr>          <chr> <dbl>   <dbl> <dbl> <dbl>        <dbl>        <dbl>
1 V0174416419V  H12   95.7   -274. -274. -95.7         178.         178.}

\item{qc_plate}{# the stats for the complete PBMC dataset - plate stats (note : this is a tibble in list)}
}
\value{
df with scores (note : Tibble is of type list)
}
\description{
Getting plate scores
}
