% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_QC_module_forShiny.R
\name{get_well_scores}
\alias{get_well_scores}
\title{Getting well scores}
\usage{
get_well_scores(df, qc_well)
}
\arguments{
\item{df}{Stat summaries for NEW PLATES - well}

\item{qc_well}{# the stats for the complete PBMC dataset - well stats}
}
\value{
df with scores
}
\description{
Getting well scores
}
