% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{heatmap_groupMaker}
\alias{heatmap_groupMaker}
\title{Title: Visualises the layout of the Seahorse plate, through different colors.}
\usage{
heatmap_groupMaker(XFe96data, flnme)
}
\arguments{
\item{XFe96data}{A Raw data sheet from a Seahorse Excel file (converted from .asyr file) which is put in a variable called XFe96data.}

\item{flnme}{Name of the imported Seahorse file.}
}
\value{
variable gg_plot which contains information about to plot the plate layout.
}
\description{
Title: Visualises the layout of the Seahorse plate, through different colors.
}
\keyword{internal}
