% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{import_excel_data}
\alias{import_excel_data}
\title{Imports an excel file.}
\usage{
import_excel_data(
  excel_data_path,
  sheet = NULL,
  range = NULL,
  col_names = TRUE,
  col_types = NULL
)
}
\arguments{
\item{excel_data_path}{Relative path, from project directory to the excel file located at the data.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the first sheet.}

\item{range}{A cell range to read from, as described in cell-specification. Includes typical Excel ranges like "B3:D87", possibly including the sheet name like "Budget!B2:G14", and more. Interpreted strictly, even if the range forces the inclusion of leading or trailing empty rows or columns. Takes precedence over skip, n_max and sheet.}

\item{col_names}{TRUE to use the first row as column names, FALSE to get default names, or a character vector giving a name for each column. If user provides col_types as a vector, col_names can have one entry per column, i.e. have the same length as col_types, or one entry per unskipped column.}

\item{col_types}{Either NULL to guess all from the spreadsheet or a character vector containing one entry per column from these options: "skip", "guess", "logical", "numeric", "date", "text" or "list". If exactly one col_type is specified, it will be recycled. The content of a cell in a skipped column is never read and that column will not appear in the data frame output. A list cell loads a column as a list of length 1 vectors, which are typed using the type guessing logic from col_types = NULL, but on a cell-by-cell basis.}
}
\value{
A data tibble if the Excel file exists, or an error or warning with the conditionaMessage.
}
\description{
Imports an excel file.
}
\examples{
out <- import_excel_data(excel_data_path, sheet = "Raw")
}
