% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{plot_BKGD_auc_facet}
\alias{plot_BKGD_auc_facet}
\title{Title}
\usage{
plot_BKGD_auc_facet(
  var,
  auc_var,
  df,
  O2_targetEmission,
  pH_targetEmission,
  file_name
)
}
\arguments{
\item{var}{}

\item{auc_var}{}

\item{df}{}

\item{O2_targetEmission}{}

\item{pH_targetEmission}{}

\item{file_name}{}

\item{AUC_bkgd}{}

\item{flnme}{}
}
\value{



}
\description{
Title
}
\examples{
Title

}
