% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{plot_raw_BKGD}
\alias{plot_raw_BKGD}
\title{Plot Raw background. Visualizes the fluorescent emission (Au) of each background well against time.}
\usage{
plot_raw_BKGD(var, total_df, O2_targetEmission, pH_targetEmission, flnme)
}
\arguments{
\item{var}{The type of Raw Emission correction, "O2_em_corr" or "pH_em_corr".}

\item{total_df}{The preprocessed Raw data sheet of the Seahorse Excel file (converted from .asyr).}

\item{O2_targetEmission}{Target emission of the O2 (Usually 12500 Fluorescence Emission (Au)).}

\item{pH_targetEmission}{Target emission of the pH (Usually 30000 Fluorescence Emission (Au)).}

\item{flnme}{Name of the imported Seahorse file.}
}
\value{
A variable gg_plot which contains information about how to plot/visualise the fluorescent emission (Au) of each background well against time.
}
\description{
Plot Raw background. Visualizes the fluorescent emission (Au) of each background well against time.
}
\keyword{internal}
