% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{plot_raw_BKGD_means}
\alias{plot_raw_BKGD_means}
\title{Title}
\usage{
plot_raw_BKGD_means(var, total_df, O2_targetEmission, pH_targetEmission, flnme)
}
\arguments{
\item{var}{}

\item{total_df}{}

\item{O2_targetEmission}{}

\item{pH_targetEmission}{}

\item{flnme}{}
}
\value{

}
\description{
Title
}
