% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_preprocess_plate_data_new.R
\name{preprocess_plate_data_2}
\alias{preprocess_plate_data_2}
\title{Preprocessing the plate data that was read in funs_read_plate_data.R}
\usage{
preprocess_plate_data_2(plate_df)
}
\arguments{
\item{plate_df}{A list of all data read from the original excel wace exported file}
}
\value{
a list with the new dataframe and the assay_info. A new dataframe called XFe96data is returned. 
The preprocessed dataframe has the following columns:
  * plate_id
  * well
  * measurement
  * tick
  * timescale
  * minutes
  * group
  * interval
  * injection
  * O2_em_corr
  * pH_em_corr
  * O2_mmHg
  * pH
  * pH_em_corr_corr
  * backgrounds for O2_em_corr, pH_em_corr, O2_mmHg, pH, pH_em_corr_corr

The assay_info has the following information:
  * ....
}
\description{
The preprocess_plate_data function preprocesses the data by:
  * changing columns names
  * adding new time columns
  * adding injection info columns
  * add plate_id column
  * calculating background data
  * calculating raw pH emission data
}
\examples{
preprocess_plate_data(plate_df)

}
