% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel_rwave}
\alias{read_excel_rwave}
\title{Title}
\usage{
read_excel_rwave(
  path,
  sheet = NULL,
  range = NULL,
  col_names = TRUE,
  col_types = NULL,
  na = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl_progress(),
  .name_repair = "unique"
)
}
\arguments{
\item{path}{}

\item{sheet}{}

\item{range}{}

\item{col_names}{}

\item{col_types}{}

\item{na}{}

\item{trim_ws}{}

\item{skip}{}

\item{n_max}{}

\item{guess_max}{}

\item{progress}{}

\item{.name_repair}{}
}
\value{

}
\description{
Title
}
