% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_preprocess_plate_data_new.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename the columns of the Raw data sheet}
\usage{
rename_columns(plate_df)
}
\arguments{
\item{plate_df}{A dataframe that was read by read_excel from the WAVE excel output file (sheet "Raw")}
}
\value{
A new dataframe with adjusted and selected column names. The selected columns with new names are:
  ("measurement","tick", "well", "group", "time", "O2_em_corr","pH_em_corr", "O2_mmHg", "pH")
}
\description{
This function is called in the preprocessing function and renames the columns of the dataframe that was
  read in by read_excel from the WAVE excel output file, excel sheet "Raw".
}
\examples{
rename_columns(XFe96data)
}
