% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_QC_module_forShiny.R
\name{background_QC_1}
\alias{background_QC_1}
\title{Title}
\usage{
background_QC_1(XFe96data_tibble, working_directory)
}
\arguments{
\item{XFe96data_tibble}{}

\item{working_directory}{}
}
\value{

}
\description{
Title
}
