% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{get_assay_info}
\alias{get_assay_info}
\title{Title : Get useful information from the Assay Configuration sheet from the Seahorse Excel file.}
\usage{
get_assay_info(filePathSeahorse)
}
\arguments{
\item{filePathSeahorse}{}
}
\value{
assayConfigurationTibble : A tibble with all the nessecary information derived from the Seahorse Assay Configuration sheet.
}
\description{
Title : Get useful information from the Assay Configuration sheet from the Seahorse Excel file.
}
