% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_QC_module_forShiny.R
\name{get_bkg_quality_colors}
\alias{get_bkg_quality_colors}
\title{Title}
\usage{
get_bkg_quality_colors(well_scores_tibble_list, plate_scores_tibble_list)
}
\arguments{
\item{well_scores_tibble_list}{}

\item{plate_scores_tibble_list}{}
}
\value{

}
\description{
Title
}
