% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_QC_module_forShiny.R
\name{get_BKGD_auc}
\alias{get_BKGD_auc}
\title{Title}
\usage{
get_BKGD_auc(total_df, var, targetEMS)
}
\arguments{
\item{total_df}{}

\item{var}{}

\item{targetEMS}{}
}
\value{

}
\description{
Title
}
