% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{get_platelayout_data}
\alias{get_platelayout_data}
\title{Title}
\usage{
get_platelayout_data(fileName, my_sheet, my_range, my_param)
}
\arguments{
\item{fileName}{}

\item{my_sheet}{}

\item{my_range}{}

\item{my_param}{}
}
\value{

}
\description{
Title
}
