% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plots.R
\name{plot_group_emissions}
\alias{plot_group_emissions}
\title{Title}
\usage{
plot_group_emissions(
  var,
  total_df,
  O2_targetEmission,
  pH_targetEmission,
  flnme
)
}
\arguments{
\item{var}{}

\item{total_df}{}

\item{O2_targetEmission}{}

\item{pH_targetEmission}{}

\item{flnme}{}
}
\value{

}
\description{
Title
}
