% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_preprocess_plate_data_new.R
\name{preprocess_plate_data_2}
\alias{preprocess_plate_data_2}
\title{Preprocessing the plate data that was read in funs_read_plate_data.R}
\usage{
preprocess_plate_data_2(plate_df)
}
\arguments{
\item{plate_df}{A tibble of all data read from the original excel wace exported file}
}
\value{
XFe96data_tibble : a tibble with the new dataframe and the assay_info. A new dataframe called XFe96data_tibble is returned.

Example XFe96data_tibble: 
#'   # A tibble: 1 × 7
assay_info contains all important assay information 

XFe96data_tibble - raw_data 
raw_data contains a list with all important raw data:
  * plate_id
  * well
  * measurement
  * tick
  * timescale
  * minutes
  * group
  * interval
  * injection
  * O2_em_corr
  * pH_em_corr
  * O2_mmHg
  * pH
  * pH_em_corr_corr
  * backgrounds for O2_em_corr, pH_em_corr, O2_mmHg, pH, pH_em_corr_corr
}
\description{
The preprocess_plate_data function preprocesses the data by:
  * changing columns names
  * adding new time columns
  * adding injection info columns
  * add plate_id column
  * calculating background data
  * calculating raw pH emission data
}
