% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel_rwave}
\alias{read_excel_rwave}
\title{Title}
\usage{
read_excel_rwave(
  path,
  sheet = NULL,
  range = NULL,
  col_names = TRUE,
  col_types = NULL,
  na = "",
  trim_ws = TRUE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  progress = readxl_progress(),
  .name_repair = "unique"
)
}
\arguments{
\item{path}{path to excel file}

\item{sheet}{sheet of specific excel file}

\item{range}{range of cellls in excel file}

\item{col_names}{column names of excel file}

\item{col_types}{column type of excel file}

\item{na}{requires description}

\item{trim_ws}{requires description}

\item{skip}{requires description}

\item{n_max}{requires description}

\item{guess_max}{requires description}

\item{progress}{requires description}

\item{.name_repair}{requires description}
}
\value{
requires description
}
\description{
Title
}
\keyword{internal}
