% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_read_plate_data.R
\name{read_plate_data}
\alias{read_plate_data}
\title{Read Plate data to return all nessecary Seahorse information.}
\usage{
read_plate_data(filePathSeahorse, injscheme)
}
\arguments{
\item{filePathSeahorse}{Absolute path to the Seahorse excel data derived from the Agilent Seahorse XF Wave software. assay result file (.asyr) files must converted into .excel files.}

\item{injscheme}{Type of injection Scheme}
}
\value{
plate_df tibble (Note: tibble has type list) with all the information important data derived from the Seahorse Excel file.
}
\description{
Read Plate data to return all nessecary Seahorse information.
}
\examples{
read_plate_data("/home/xiang/Documents/Documents Gerwin/Projects/r_exersice/wur_package_exersice/data/raw_data/20191219_SciRep_PBMCs_donor_A.xlsx", "HAP")
}
